requirements.txt
streamlit
PyPDF2
pytesseract
pdf2image
langchain
langchain-google-genai
googletrans==4.0.0-rc1
faiss-cpu
python-dotenv
Pillow
